// kservices_ts_api/src/api/domain/entities/v3/ivr.entity.ts
export type setLiveCallsRequestValidate = {
  sme_id: number;
  date_time: string;
  longcode: string;
  customer_number: string;
  agent_number: string;
  self_ip: string;
  agent_id: number;
  session_id: string;
  call_status: number;
  call_type: string;
  is_auto_dial: number;
  customer_name: string;
};

export type updateLiveCallsRequestValidate = {
  sme_id: number;
  agent_number: string;
  agent_id: number;
  session_id: string;
  call_status: number;
  customer_number: string;
};

export type deleteLiveCallsValidate = {
  session_id: string;
  sme_id: number;
};
// ...
export type setUniqueCallsRequestValidate = {
  sme_id: number;
  channel: string;
  callpatchedAgentGroup: string;
  callDirection: string;
  callDirectionStatus: string;
  callRecordedFile: string;
  callRecordingStatus: string;
  customerNumber: string;
  agentNumber: string;
  callcdrMode: string;
  callMode: string;
  callchannelNo: string;
  callDuration: string;
  callendDateTime: string;
  callHlr: string;
  callinsertDateTime: string;
  calllongcode: string;
  callmasterShortcode: string;
  callpatchedAgentId: string;
  ip: string;
  callshortcodeMapping: string;

  callsmeIdentifier: string;
  callstartDateTime: string;
  voicemailRecordingFile: string;
  voicemailRecordingStatus: string;
  callgroupid: string;
  callSessionId: string;
  callcdrFlag: string;
  callanswerFlag: string;
  callfinalStatus: string;
  calldisconnectedBy: string;
  addressBookName: string;
  addressBookId: string;
  insertDateTime: string;
  provisionalFlag: number;
  agentName: string;
};

export type setblacklistRequestValidate = {
  sme_id: number;
  start_date: string;
  longcode: string;
  customer_number: string;
  session_id: string;
  insertDateTime: string;
};

export type setKeepAliveRequest = {
  start_date: string;
};